Type formula helpers for fractional effective reps
RPE now feeds a fractional RIR into effective reps (RPE 8.5 = 1.5 RIR), so
the per-formula helpers took a float all along despite being typed `r: int`.
Fixed the annotations and added a test pinning the 12-rep curvilinear cutoff
at a fractional boundary (12.5 drops Brzycki/Lander/Mayhew, exactly 12.0
keeps them), in both engines.

+++ src/liftmath/onerm.py
@@ -49,27 +49,27 @@ import math
 from dataclasses import dataclass, field
 
 
-def _epley(w: float, r: int) -> float:
+def _epley(w: float, r: float) -> float:
     return w * (1 + r / 30.0)
 
 
-def _brzycki(w: float, r: int) -> float:
+def _brzycki(w: float, r: float) -> float:
     return w * 36.0 / (37.0 - r) if r < 37 else float("nan")
 
 
-def _lombardi(w: float, r: int) -> float:
+def _lombardi(w: float, r: float) -> float:
     return w * (r ** 0.10)
 
 
-def _oconner(w: float, r: int) -> float:
+def _oconner(w: float, r: float) -> float:
     return w * (1 + 0.025 * r)
 
 
-def _lander(w: float, r: int) -> float:
+def _lander(w: float, r: float) -> float:
     return 100.0 * w / (101.3 - 2.67123 * r)
 
 
-def _mayhew(w: float, r: int) -> float:
+def _mayhew(w: float, r: float) -> float:
     return 100.0 * w / (52.2 + 41.9 * math.exp(-0.055 * r))
